added unSeen count

+++ back-end/src/main/java/com/example/demo/dtos/ChatDTO.java
@@ -12,7 +12,7 @@ import lombok.NoArgsConstructor;
 public class ChatDTO {
     private String username;
     private String email;
-    private int unreadCount;
+    private Long unreadCount;
     private String preview;
     private String title;
